Removes usage of ConcurrentTaskScheduler
ConcurrentTaskScheduler doesn't expose any way for it's inner executor to be shutdown,
when it's instance is created with the default constructor. This can lead to a scenario
where a library user ends up with many unused threads with no clear cues about where they
came from.

This commit replaces the usage of ConcurrentTaskScheduler with ThreadPoolTaskScheduler,
a type of similar semantics but better API, which includes methods for managing its
internal resources lifecycle. Also it implements spring-beans interfaces which makes
it more suitable to be managed by a Spring application context.

- Fixes #624