Override Java 1.8 default Map methods in ObservableMap
This is to ensure that default map methods will be passed to the delegate, which is in this case an instance of ConcurrentHashMap.

Without this change, an attempt to modify state variables via default methods will be delegated to default Map implementation, which is not thread safe. These methods are already overridden in ConcurrentHashMap.